fix: set right error message on customization & refactoring of prop drilling
