+++ src/components/Chat/Chat.tsx
@@ -6,6 +6,7 @@ import {
   useAppSelector,
   useAppDispatch,
   useSendChatRequest,
+  useGetPromptsQuery,
 } from "../../hooks";
 import type { Config } from "../../features/Config/configSlice";
 import {
@@ -18,9 +19,12 @@ import {
   selectChatId,
   selectMessages,
   getSelectedToolUse,
+  getSelectedSystemPrompt,
+  setSystemPrompt,
 } from "../../features/Chat/Thread";
 import { ThreadHistoryButton } from "../Buttons";
 import { push } from "../../features/Pages/pagesSlice";
+import { SystemPrompts } from "../../services/refact";
 
 export type ChatProps = {
   host: Config["host"];
@@ -31,9 +35,6 @@ export type ChatProps = {
   // TODO: update this
   caps: ChatFormProps["caps"];
   maybeSendToSidebar: ChatFormProps["onClose"];
-  prompts: ChatFormProps["prompts"];
-  onSetSystemPrompt: ChatFormProps["onSetSystemPrompt"];
-  selectedSystemPrompt: ChatFormProps["selectedSystemPrompt"];
 };
 
 export const Chat: React.FC<ChatProps> = ({
@@ -41,9 +42,6 @@ export const Chat: React.FC<ChatProps> = ({
   unCalledTools,
   caps,
   maybeSendToSidebar,
-  prompts,
-  onSetSystemPrompt,
-  selectedSystemPrompt,
 }) => {
   const [isViewingRawJSON, setIsViewingRawJSON] = useState(false);
   const chatContentRef = useRef<HTMLDivElement>(null);
@@ -57,6 +55,10 @@ export const Chat: React.FC<ChatProps> = ({
   const dispatch = useAppDispatch();
   const messages = useAppSelector(selectMessages);
 
+  const promptsRequest = useGetPromptsQuery();
+  const selectedSystemPrompt = useAppSelector(getSelectedSystemPrompt);
+  const onSetSelectedSystemPrompt = (prompt: SystemPrompts) =>
+    dispatch(setSystemPrompt(prompt));
   const [isDebugChatHistoryVisible, setIsDebugChatHistoryVisible] =
     useState(false);
 
@@ -148,8 +150,8 @@ export const Chat: React.FC<ChatProps> = ({
         onStopStreaming={abort}
         onClose={maybeSendToSidebar}
         onTextAreaHeightChange={onTextAreaHeightChange}
-        prompts={prompts}
-        onSetSystemPrompt={onSetSystemPrompt}
+        prompts={promptsRequest.data ?? {}}
+        onSetSystemPrompt={onSetSelectedSystemPrompt}
         selectedSystemPrompt={selectedSystemPrompt}
       />
       <Flex justify="between" pl="1" pr="1" pt="1">
